requirements.txt
amqp==2.3.2
argon2-cffi==18.1.0
bcrypt==3.1.4
billiard==3.5.0.3
celery==4.1.1
certifi==2019.3.9
cffi==1.11.5
click==6.7
coverage==4.5.1
cycler==0.10.0
Django==2.2.1
django-bootstrap3==10.0.1
django-braces==1.13.0
django-crispy-forms==1.8.1
django-debug-toolbar==1.9.1
django-material==1.5.6
Faker==1.0.1
Flask==0.12.2
Flask-Login==0.4.1
Flask-Script==2.0.6
gunicorn==19.9.0
itsdangerous==0.24
Jinja2==2.10
kiwisolver==1.1.0
kombu==4.2.1
MarkupSafe==1.0
matplotlib==3.0.3
misaka==2.1.0
mysqlclient==1.3.12
nltk==3.4
numpy==1.13.3
pandas==0.24.2
passlib==1.7.1
Pillow==5.1.0
pipenv==2018.11.26
psycopg2==2.7.4
pycparser==2.18
pyparsing==2.4.0
python-dateutil==2.7.5
pytz==2017.3
singledispatch==3.4.0.3
six==1.11.0
sqlparse==0.2.4
text-unidecode==1.2
vine==1.1.4
virtualenv==15.1.0
virtualenv-clone==0.5.3
virtualenvwrapper-win==1.2.5
Werkzeug==0.13
